requirements.txt
fastapi
uvicorn
lancedb
ollama
llama-index-core
llama-index
pyarrow
pyyaml
mcp[cli]
langdetect
torch
sentence-transformers
